template param list does not need to match function list, but types need to be match

+++ 10_Templates/UnitsLibrary/Solution/Main.cc
@@ -4,12 +4,12 @@
 
 int main()
 {
-    const float x = 5.0F;
+    const double x = 5.0F;
     const float y = 4.0F;
 
     const auto deg = 10.0F;
     const auto rad = deg_to_rad(deg);
-
+	
     const auto [x_, y_] = rotate_point(rad, x, y);
     std::cout << x_ << " " << y_ << std::endl;
